backend/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.9
redis==4.6.0
django-redis==5.4.0
celery==5.3.6
qrcode==7.4.2
Pillow==10.0.0
django-celery-beat==2.5.0
drf-yasg==1.21.7
python-decouple==3.8
django-filter==23.5